API configuration ventas and detalle ventas

+++ src/main/java/co/edu/unbosque/ciclo3back/api/detalle_ventasAPI.java
@@ -2,6 +2,10 @@ package co.edu.unbosque.ciclo3back.api;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,32 +15,73 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import co.edu.unbosque.ciclo3back.dao.detalle_ventasDAO;
+import co.edu.unbosque.ciclo3back.dao.mensaje;
 import co.edu.unbosque.ciclo3back.model.detalle_ventas;
+import co.edu.unbosque.ciclo3back.model.productos;
+
 
 
 @RestController 
-@RequestMapping("detalle_ventas")
+@RequestMapping("detalleVenta")
 public class detalle_ventasAPI {
 	@Autowired 
 	private detalle_ventasDAO detalle_ventasDAO;
 
-	@PostMapping("/guardar")
-	public void guardar(@RequestBody detalle_ventas detalle_ventas) {
-		detalle_ventasDAO.save(detalle_ventas);
-	}
-
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
 	@GetMapping("/listar")
 	public List<detalle_ventas> listar() {
 		return detalle_ventasDAO.findAll();
 	}
-
-	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id") Integer id) {
-		detalle_ventasDAO.deleteById(id);
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@PostMapping("/guardar")
+	public ResponseEntity<?> guardar( @RequestBody detalle_ventas detalle_venta, BindingResult bindingResult) {
+		if(bindingResult.hasErrors())
+            return new ResponseEntity(new mensaje("Datos mal ingresados"), HttpStatus.BAD_REQUEST);
+        if(detalle_ventasDAO.existsById(detalle_venta.getCodigo_detalle_venta()))
+            return new ResponseEntity(new mensaje("Ya existe un detalle de venta con el codigo ingresado"), HttpStatus.BAD_REQUEST);
+        detalle_ventasDAO.save(detalle_venta);
+        return new ResponseEntity(new mensaje("Detalle venta agregada con exito"), HttpStatus.CREATED);
+	}
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@PutMapping("/actualizar/{id}")
+	public ResponseEntity<?> actualizar(@RequestBody detalle_ventas detalle_venta, BindingResult bindingResult, @PathVariable("id") Long id) {
+		if(bindingResult.hasErrors())
+            return new ResponseEntity(new mensaje("Datos mal ingresados"), HttpStatus.BAD_REQUEST);
+        if(!detalle_ventasDAO.existsById(id))
+            return new ResponseEntity(new mensaje("No existe el detalle de venta a actualizar"), HttpStatus.NOT_FOUND);
+        detalle_ventas detalleActualizar = detalle_ventasDAO.findById(id).get();
+        detalleActualizar.setCodigo_detalle_venta(detalle_venta.getCodigo_detalle_venta());
+        detalleActualizar.setCantidad_producto(detalle_venta.getCantidad_producto());
+        detalleActualizar.setCodigo_producto(detalle_venta.getCodigo_producto());
+        detalleActualizar.setCodigo_venta(detalle_venta.getCodigo_venta());
+        detalleActualizar.setValor_total(detalle_venta.getValor_total());
+        detalleActualizar.setValor_venta(detalle_venta.getValor_venta());
+        detalleActualizar.setValoriva(detalle_venta.getValoriva());
+        detalle_ventasDAO.save(detalleActualizar);
+        return new ResponseEntity(new mensaje("Detalle venta actualizada"), HttpStatus.OK);
 	}
 
-	@PutMapping("/actualizar")
-	public void actualizar(@RequestBody detalle_ventas detalle_ventas) {
-		detalle_ventasDAO.save(detalle_ventas);
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@GetMapping("/detalle/{id}")
+	public ResponseEntity<productos> consultar(@PathVariable("id") Long id){
+		if(!detalle_ventasDAO.existsById(id)) 
+			return new ResponseEntity(new mensaje("No existe el detalle venta"), HttpStatus.NOT_FOUND);
+		detalle_ventas detalle = detalle_ventasDAO.findById(id).get();
+		return new ResponseEntity(detalle, HttpStatus.OK);
+	}
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@DeleteMapping("/eliminar/{id}")
+	public ResponseEntity<?> eliminar(@PathVariable("id") Long id) {
+		if(!detalle_ventasDAO.existsById(id))
+			return new ResponseEntity(new mensaje("No existe el detalle venta"), HttpStatus.NOT_FOUND);
+		detalle_ventasDAO.deleteById(id);
+		return new ResponseEntity(new mensaje("Detalle venta eliminada"), HttpStatus.OK);
 	}
 }
