+++ src/main/java/co/edu/unbosque/ciclo3back/api/ventasAPI.java
@@ -2,6 +2,10 @@ package co.edu.unbosque.ciclo3back.api;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,7 +14,10 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import co.edu.unbosque.ciclo3back.dao.mensaje;
 import co.edu.unbosque.ciclo3back.dao.ventasDAO;
+import co.edu.unbosque.ciclo3back.model.productos;
 import co.edu.unbosque.ciclo3back.model.ventas;
 
 
@@ -20,23 +27,60 @@ public class ventasAPI {
 	@Autowired 
 	private ventasDAO ventasDAO;
 
-	@PostMapping("/guardar")
-	public void guardar(@RequestBody ventas ventas) {
-		ventasDAO.save(ventas);
-	}
-
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
 	@GetMapping("/listar")
 	public List<ventas> listar() {
 		return ventasDAO.findAll();
 	}
-
-	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id") Integer id) {
-		ventasDAO.deleteById(id);
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@PostMapping("/guardar")
+	public ResponseEntity<?> guardar( @RequestBody ventas venta, BindingResult bindingResult) {
+		if(bindingResult.hasErrors())
+            return new ResponseEntity(new mensaje("Datos mal ingresados"), HttpStatus.BAD_REQUEST);
+        if(ventasDAO.existsById(venta.getCodigo_venta()))
+            return new ResponseEntity(new mensaje("Ya existe una venta con el codigo ingresado"), HttpStatus.BAD_REQUEST);
+        ventasDAO.save(venta);
+        return new ResponseEntity(new mensaje("Venta agregado con exito"), HttpStatus.CREATED);
+	}
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@PutMapping("/actualizar/{id}")
+	public ResponseEntity<?> actualizar(@RequestBody ventas venta, BindingResult bindingResult, @PathVariable("id") Long id) {
+		if(bindingResult.hasErrors())
+            return new ResponseEntity(new mensaje("Datos mal ingresados"), HttpStatus.BAD_REQUEST);
+        if(!ventasDAO.existsById(id))
+            return new ResponseEntity(new mensaje("No existe la venta a actualizar"), HttpStatus.NOT_FOUND);
+        ventas ventaActualizar = ventasDAO.findById(id).get();
+        ventaActualizar.setCodigo_venta(venta.getCodigo_venta());
+        ventaActualizar.setCedula_cliente(venta.getCedula_cliente());
+        ventaActualizar.setCedula_usuario(venta.getCedula_usuario());
+        ventaActualizar.setIvaventa(venta.getIvaventa());
+        ventaActualizar.setTotal_venta(venta.getTotal_venta());
+        ventaActualizar.setValor_venta(venta.getValor_venta());
+        ventasDAO.save(ventaActualizar);
+        return new ResponseEntity(new mensaje("Venta actualizada"), HttpStatus.OK);
 	}
 
-	@PutMapping("/actualizar")
-	public void actualizar(@RequestBody ventas ventas) {
-		ventasDAO.save(ventas);
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@GetMapping("/detalle/{id}")
+	public ResponseEntity<productos> consultar(@PathVariable("id") Long id){
+		if(!ventasDAO.existsById(id)) 
+			return new ResponseEntity(new mensaje("No existe la venta"), HttpStatus.NOT_FOUND);
+		ventas venta = ventasDAO.findById(id).get();
+		return new ResponseEntity(venta, HttpStatus.OK);
+	}
+	
+	@CrossOrigin(origins = {"http://localhost:3000", "https://ciclo3-mintic-front.herokuapp.com"})
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@DeleteMapping("/eliminar/{id}")
+	public ResponseEntity<?> eliminar(@PathVariable("id") Long id) {
+		if(!ventasDAO.existsById(id))
+			return new ResponseEntity(new mensaje("No existe la venta"), HttpStatus.NOT_FOUND);
+		ventasDAO.deleteById(id);
+		return new ResponseEntity(new mensaje("Venta eliminada"), HttpStatus.OK);
 	}
 }
